feat: mysite03-BoardController finish

+++ mysite03/src/main/java/com/poscodx/mysite/controller/BoardController.java
@@ -35,8 +35,8 @@ public class BoardController {
 		return "board/list";
 	}
 	
-	@RequestMapping(value="/write/{pageNum}", method=RequestMethod.GET) // write form
-	public String write(@PathVariable("pageNum") String pageNum,
+	@RequestMapping(value="/writeform/{no}/{pageNum}", method=RequestMethod.GET) // write form
+	public String write(@PathVariable("no") String no, @PathVariable("pageNum") String pageNum,
 			HttpSession session, Model model) {
 		
 		// Access Control(접근 제어) 
@@ -44,14 +44,14 @@ public class BoardController {
 		if(authUser == null) {
 			return "redirect:/user/noSession";
 		}
-		model.addAttribute("pageNum", pageNum);
 		
 		return "board/write";
 	}
-	
+
 	@RequestMapping(value="/write/{no}/{pageNum}", method=RequestMethod.POST) // write 
-	public String write(@PathVariable("no") String no, 
-			@PathVariable("pageNum") String pageNum,
+	public String write(@PathVariable("no") Long no, @PathVariable("pageNum") String pageNum,
+			@RequestParam(value="title", required=true, defaultValue="") String title,
+			@RequestParam(value="content", required=true, defaultValue="") String content,
 			HttpSession session, Model model) {
 		
 		// Access Control(접근 제어) 
@@ -60,22 +60,48 @@ public class BoardController {
 			return "redirect:/user/noSession";
 		}
 		
-		System.out.println("no : " + no + "pageNum : " + pageNum);
-		model.addAttribute("pageNum", pageNum);
+		BoardVo vo = boardService.getBoardVoForWrite(no, authUser, title, content);
+		boardService.insert(vo);
 		
-		return "board/write";
+		return "redirect:/board/" + pageNum;	
+	}
+	
+	@RequestMapping(value="/viewform/{no}/{pageNum}", method=RequestMethod.GET) // view form
+	public String write(@PathVariable("no") Long no, @PathVariable("pageNum") String pageNum, 
+			@RequestParam(value="title", required=true, defaultValue="") String title,
+			@RequestParam(value="content", required=true, defaultValue="") String content,
+			Model model) {
+		
+		BoardVo boardVo = boardService.getPost(no);
+		model.addAttribute("vo", boardVo);
+		
+		return "board/view";
 	}
 	
 	@RequestMapping(value="/deleteform/{no}/{pageNum}", method=RequestMethod.GET) // delete form
-	public String delete(@PathVariable("no") Long no, 
-			@PathVariable("pageNum") String pageNum, Model model) {
-		model.addAttribute("no", no);
+	public String delete(@PathVariable("no") Long no, @PathVariable("pageNum") String pageNum, 
+			Model model, HttpSession session) {
+		
+		// Access Control(접근 제어) 
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/noSession";
+		}
+
 		return "board/delete";
 	}
 
 	@RequestMapping(value="/delete/{no}/{pageNum}", method=RequestMethod.POST) // delete action
 	public String delete(@PathVariable("no") Long no, @PathVariable("pageNum") String pageNum,
-			@RequestParam(value="password", required=true, defaultValue="") String password) {
+			@RequestParam(value="password", required=true, defaultValue="") String password,
+			HttpSession session) {
+		
+		// Access Control(접근 제어) 
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/noSession";
+		}
+				
 		boolean result = boardService.deletePost(no, password);
 	
 		if(result) { // 비밀번호 일치 - 삭제 성공 
@@ -83,5 +109,36 @@ public class BoardController {
 		} else { // 비밀번호불일치 - 삭제 실패 
 			return "redirect:/board/deleteform/" + no + "/" + pageNum;
 		}	
+	}
+	
+	@RequestMapping("/modifyform/{no}/{pageNum}")
+	public String update(HttpSession session, Model model,
+			@PathVariable("no") Long no, @PathVariable("pageNum") String pageNum) {
+		// Access Control(접근 제어) 
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/noSession";
+		}
+		///////////////////////////////////////////////////
+		
+		BoardVo boardVo = boardService.getPost(no); 
+		model.addAttribute("vo", boardVo);
+		
+		return "board/modify";
+	}
+	
+	@RequestMapping(value="/modify/{no}/{pageNum}", method=RequestMethod.POST)
+	public String update(HttpSession session, BoardVo boardVo,
+			@PathVariable("no") Long no, @PathVariable("pageNum") String pageNum) {
+		// Access Control(접근 제어) 
+		UserVo authUser = (UserVo) session.getAttribute("authUser");
+		if(authUser == null) {
+			return "redirect:/user/noSession";
+		}
+		///////////////////////////////////////////////////
+		
+		boardService.update(no, boardVo);
+		
+		return "redirect:/board/viewform/" + no + "/" + pageNum;
 	}
 }
